chore: move shouldUseContrastColor to general utils dir
