+++ src/visualizations/config/adapters/dhis_highcharts/customSVGOptions/singleValue/config/getSingleValueTextColor.js
@@ -1,7 +1,7 @@
 import { colors } from '@dhis2/ui'
 import { LEGEND_DISPLAY_STYLE_TEXT } from '../../../../../../../modules/legends.js'
+import { shouldUseContrastColor } from '../../../../../../util/shouldUseContrastColor.js'
 import { getSingleValueLegendColor } from './getSingleValueLegendColor.js'
-import { shouldUseContrastColor } from './shouldUseContrastColor.js'
 
 export function getSingleValueTextColor(
     baseColor,
