+++ src/visualizations/util/shouldUseContrastColor.js
